Replacing some signed integers with unsigned integers.

+++ zodiac_lib/src/atom.cpp
@@ -7,10 +7,10 @@ namespace Zodiac
 {
 
 static void atom_table_grow(Atom_Table *at);
-static void atom_table_add_block(Atom_Table *at, Atom_Block *new_block, i64 block_size);
-static Atom atom_table_add(Atom_Table *at, const char *cstr, i64 length);
+static void atom_table_add_block(Atom_Table *at, Atom_Block *new_block, u64 block_size);
+static Atom atom_table_add(Atom_Table *at, const char *cstr, u64 length);
 
-void atom_table_init(Allocator *allocator, Atom_Table *at, i64 initial_capacity/*= ATOM_TABLE_INITIAL_CAPACITY*/)
+void atom_table_init(Allocator *allocator, Atom_Table *at, u64 initial_capacity/*= ATOM_TABLE_INITIAL_CAPACITY*/)
 {
     at->allocator = allocator;
 
@@ -23,7 +23,7 @@ void atom_table_init(Allocator *allocator, Atom_Table *at, i64 initial_capacity/
     atom_table_add_block(at, &at->first_block, ATOM_TABLE_INITIAL_BLOCK_SIZE);
 }
 
-void atom_table_init(Atom_Table *at, i64 initial_capacity /*= ATOM_TABLE_INITIAL_CAPACITY*/)
+void atom_table_init(Atom_Table *at, u64 initial_capacity /*= ATOM_TABLE_INITIAL_CAPACITY*/)
 {
     atom_table_init(&dynamic_allocator, at, initial_capacity);
 }
@@ -62,12 +62,12 @@ static void atom_table_grow(Atom_Table *at)
     at->atoms = new_atoms;
     at->hashes = new_hashes;
 
-    for (i64 i = 0; i < old_cap; i++) {
+    for (u64 i = 0; i < old_cap; i++) {
 
         u64 hash = old_hashes[i];
-        i64 hash_index = (i64)(hash % (u64)at->capacity);
+        u64 hash_index = hash % at->capacity;
 
-        i64 iteration_count = 0;
+        u64 iteration_count = 0;
 
         while (iteration_count < at->capacity) {
 
@@ -90,7 +90,7 @@ static void atom_table_grow(Atom_Table *at)
     free(at->allocator, old_atoms);
 }
 
-static void atom_table_add_block(Atom_Table *at, Atom_Block *new_block, i64 block_size)
+static void atom_table_add_block(Atom_Table *at, Atom_Block *new_block, u64 block_size)
 {
     assert(new_block->first == nullptr);
     assert(new_block->current == nullptr);
@@ -107,7 +107,7 @@ static void atom_table_add_block(Atom_Table *at, Atom_Block *new_block, i64 bloc
     at->current_block = new_block;
 }
 
-static Atom atom_table_add(Atom_Table *at, const char *cstr, i64 length)
+static Atom atom_table_add(Atom_Table *at, const char *cstr, u64 length)
 {
     assert(at->current_block);
 
@@ -140,12 +140,12 @@ static Atom atom_table_add(Atom_Table *at, const char *cstr, i64 length)
     return result;
 }
 
-Atom atom_get(Atom_Table *at, const char *cstr, i64 length)
+Atom atom_get(Atom_Table *at, const char *cstr, u64 length)
 {
     u64 hash = hash_c_string(cstr, length);
-    i64 hash_index = (i64)(hash % (u64)at->capacity);
+    u64 hash_index = hash % at->capacity;
 
-    i64 iteration_count = 0;
+    u64 iteration_count = 0;
 
     while (iteration_count < at->capacity) {
 
@@ -176,7 +176,7 @@ Atom atom_get(Atom_Table *at, const char *cstr, i64 length)
 
 Atom atom_get(Atom_Table *at, const char *cstr)
 {
-    auto length = (i64)strlen(cstr);
+    auto length = strlen(cstr);
     return atom_get(at, cstr, length);
 }
